requirements.txt
black==24.10.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.3.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
isort==5.13.2
numpy==2.1.3
openai==1.56.1
pillow==11.0.0
psycopg2==2.9.10
requests==2.32.3
twilio==9.3.7